check to find if internet is reachable
I added a simple decorator as well as checking for internet connection with ping to see if whether the host is down or internet is no reachable. Optionally, @retry keyword argument can be used to retry to reach internet in case it fails and @opt_exception to raise an optional exception to check for it somewhere else.